Adjust test case for xdfm to use int64 tensor instead of int32.

+++ tests/test_xdfm.py
@@ -13,7 +13,7 @@ def test_cin():
 
 
 def test_xdfm():
-    x = tf.convert_to_tensor([[1, 2, 2, 1, 5], [0, 3, 2, 2, 4], [2, 2, 2, 1, 1]])
+    x = tf.convert_to_tensor([[1, 2, 2, 1, 5], [0, 3, 2, 2, 4], [2, 2, 2, 1, 1]], dtype="int64")
     m = ExtremeDeepFactorizationMachine([3, 4, 5, 4, 6], 4, [20, 10, 1], [10, 5], .1, False)
     o = m(x)
